refactor(dice): split cog into multiple files
