+++ cogs5e/dice/__init__.py
@@ -1,0 +1,5 @@
+from .cog import Dice
+
+
+def setup(bot):
+    bot.add_cog(Dice(bot))
